fix: cancel modal
